# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the UnderNET Channel Service. It provides a RESTful interface for managing users, channels, and other resources. \u003c!-- ReDoc-Inject: \u003csecurity-definitions\u003e --\u003e\n# Authorization\n**JWT Bearer Token:** The main authorization method for the API. Needs `Authorization: Bearer \u003cjwt-token\u003e` HTTP header to authenticate.\n\u003c!-- ReDoc-Inject: \u003csecurity-definitions\u003e --\u003e",
        "title": "UnderNET Channel Service API",
        "contact": {
            "name": "Ratler",
            "email": "ratler@undernet.org"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/UndernetIRC/cservice-api/blob/master/LICENSE"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/activate": {
            "post": {
                "description": "Activates a user account using the provided token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Activate user account",
                "parameters": [
                    {
                        "description": "Activate account request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UserRegisterActivateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.UserRegisterActivateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "Creates a new user account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "Register request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UserRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.UserRegisterActivateRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string",
                    "x-order": "0"
                }
            }
        },
        "controllers.UserRegisterActivateResponse": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string",
                    "x-order": "0"
                },
                "email": {
                    "type": "string",
                    "x-order": "1"
                }
            }
        },
        "controllers.UserRegisterRequest": {
            "type": "object",
            "required": [
                "aup",
                "confirm_password",
                "coppa",
                "email",
                "password",
                "username"
            ],
            "properties": {
                "username": {
                    "type": "string",
                    "maxLength": 12,
                    "minLength": 2,
                    "x-order": "0"
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 10,
                    "x-order": "1"
                },
                "confirm_password": {
                    "type": "string",
                    "x-order": "2"
                },
                "email": {
                    "type": "string",
                    "x-order": "3"
                },
                "aup": {
                    "type": "boolean",
                    "x-order": "4"
                },
                "coppa": {
                    "type": "boolean",
                    "x-order": "5"
                }
            }
        },
...
        "errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "description": "Error code constant",
                            "type": "string"
                        },
                        "details": {
                            "description": "Additional error details (validation errors, etc.)"
                        },
                        "message": {
                            "description": "Human-readable message",
                            "type": "string"
                        }
                    }
                },
                "status": {
                    "description": "Always \"error\"",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWTBearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}